Updated the MobileNUmberPage
Now the dart file hosts the Phone Number Page and the OTP Verification Page. Once the user clicks on the Continue button, they will be redirected to the OTP Verification Screen. If it is an Android device, the system will automatically read the OTP and log in the user. For other devices, it will show an entry field and the user can then enter the OTP. It will be verified and the user will be logged in. This will lead to the Profile Page.